requirements.txt
streamlit>=1.24.0
python-dotenv>=0.19.0
oracledb>=1.3.0
openai>=1.0.0
pandas>=1.5.0
numpy>=1.23.0
pyvis>=0.3.0
networkx>=2.8.0
pypgx>=0.1.0  # Oracle Property Graph Python客户端